# api/openapi.yaml
openapi: 3.1.0
info:
  title: MyShopEdge API
  version: 0.3.0
  description: "The MyShopEdge REST API. Bookkeeping and finance for UK TikTok Shop sellers.\n\nThis document\
    \ covers the seventeen endpoints that can be built without validated\nTikTok data, plus the shared\
    \ components every other endpoint will use. The\nconventions are those already set in MSE-API-001\
    \ section 4.1 and are restated here\nas schemas so they are enforced rather than described.\n\nFour\
    \ rules govern every response in this document.\n\n1. **Money is never a decimal.** Every amount is\
    \ an integer of minor units with its\n   currency beside it. A figure of 20.84 is `{\"amount_minor\"\
    : 2084, \"currency\": \"GBP\"}`.\n   Pence do not round, and a float would make them.\n\n2. **Instants\
    \ are UTC, periods are Europe/London.** A timestamp is ISO 8601 with a Z.\n   A day or a month is\
    \ a local date, because a seller's day is a London day and a VAT\n   quarter is a London quarter.\n\
    \n3. **A figure carries its own confidence.** Any response holding a money figure the\n   seller will\
    \ act on states its basis, its period and its cost coverage, and says\n   whether it is confirmed,\
    \ estimated or incomplete. A figure that cannot be trusted\n   returns null with a reason rather than\
    \ a zero.\n\n4. **Errors are problem details with a stable code.** The `code` field is the\n   contract.\
...
    \ the Go Live review's second ground.\n"
  contact:
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /admin/accounts:
    get:
      tags:
      - Admin
      summary: Every account
      description: Every account, newest first, with its plan and number of shops. `email` narrows the list
        to the one account that holds that address, compared without regard to case, which is how an admin
        finds the account named in a data protection request.
      operationId: listAdminAccounts
      parameters:
      - name: email
        in: query
        required: false
        schema:
          type: string
          maxLength: 320
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccountList'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: &id001
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/shops:
    get:
      tags:
      - Admin
      summary: Every shop and its connection
      description: Every shop with its account, its connection state, its last sync, the latest token refresh
        failure, and how many of its statements reconcile and how many do not. No TikTok token is returned,
        and no money figure (A34.8, ruling 4).
      operationId: listAdminShops
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminShopList'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
...
  /admin/notice-emails:
    get:
      tags:
      - Admin
      summary: Notice emails by state, and the ones not yet sent
      description: How many notices are in each email state, and the notices whose email is pending or has
        failed, with Resend's last answer.
      operationId: getAdminNoticeEmails
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNoticeEmails'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/audit:
    get:
      tags:
      - Admin
      summary: The latest audit log entries
      description: The latest audit log entries across every account, newest first. Every admin action writes
        one; views write none.
      operationId: listAdminAudit
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuditList'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/figures:
    get:
      tags:
      - Admin
      summary: Sign-ups, trials and plans
      description: Sign-ups by month in Europe/London, and per plan the accounts on trial, paying and with
        a failed payment. Revenue is each paying plan's list price each month, not money taken, because
        Stripe holds the money actually taken (A34.5).
      operationId: getAdminFigures
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFigures'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
  /admin/accounts/{accountId}/suspend:
    post:
      tags:
      - Admin
      summary: Suspend an account
      description: Suspends an active account (A34.8, rulings 2 and 3). Renewal is stopped at Stripe first,
        so a Stripe failure changes nothing. The seller is then refused on every route but getMe, the daily
        sync skips the account, and a critical `account_suspended` notice is written and emailed. Writes
        one audit log entry.
      operationId: suspendAccount
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatusOut'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '409':
          description: The account is not active
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '502':
          description: Stripe could not be reached, so nothing was changed
          content: *id001
  /admin/accounts/{accountId}/reactivate:
    post:
      tags:
      - Admin
      summary: Reactivate a suspended account
      description: Reactivates a suspended account. Renewal is turned back on only where the suspension
        stopped it, and an `account_reactivated` notice is written and emailed. Writes one audit log entry.
      operationId: reactivateAccount
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatusOut'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '409':
          description: The account is not suspended
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '502':
          description: Stripe could not be reached, so nothing was changed
          content: *id001
  /admin/shops/{shopId}/sync:
    post:
      tags:
      - Admin
      summary: Sync one shop now
      description: Refreshes and syncs one shop in the background, as the daily run does. Writes one audit
        log entry.
      operationId: syncShopNow
      parameters:
      - name: shopId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedOut'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/notifications/{notificationId}/retry-email:
    post:
      tags:
      - Admin
      summary: Try a failed notice email again
      description: Puts a notice whose email failed back to pending with no attempts, and runs the account's
        email sweep in the background. Writes one audit log entry.
      operationId: retryNoticeEmail
      parameters:
      - name: notificationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetryOut'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '409':
          description: That notice's email has not failed
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/accounts/{accountId}/export:
    post:
      tags:
      - Admin
      summary: Build an account's data export on its seller's request
      description: Builds an account's data export on its seller's request, as requestAccountExport does
        for the seller. Writes one audit log entry.
      operationId: requestAccountExportForAccount
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '409':
          description: An export of this account is already being prepared
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/accounts/{accountId}/export/{exportId}:
    get:
      tags:
      - Admin
      summary: An account export's status and signed link
      description: An account export's status, and its signed link once ready, as getAccountExport gives
        the seller.
      operationId: getAccountExportForAccount
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: exportId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /admin/accounts/{accountId}/deletion:
    post:
      tags:
      - Admin
      summary: Close an account for deletion on its seller's request
      description: Closes an account for deletion on its seller's request, with the thirty days A30.1 sets,
        exactly as deleteMe does. A suspended account is refused with 409 and must be reactivated first.
        Writes one audit log entry.
      operationId: deleteAccountForSeller
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionAcknowledgement'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '409':
          description: The account is suspended
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '502':
          description: Stripe could not be reached, so nothing was changed
          content: *id001
  /admin/accounts/{accountId}/deletion/cancel:
    post:
      tags:
      - Admin
      summary: Cancel an account's deletion on its seller's request
      description: Cancels a deletion within the thirty days, on its seller's request, exactly as cancelAccountDeletion
        does. Writes one audit log entry.
      operationId: cancelAccountDeletionForSeller
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelOut'
        '404':
          description: Not an admin, not signed in, or no such account, shop, notice or export. A refused
            admin request is answered exactly as a path that does not exist.
          content: *id001
        '409':
          description: The account has no deletion to cancel
          content: *id001
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '502':
          description: Stripe could not be reached, so nothing was changed
          content: *id001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'A JWT from the identity provider, verified against its JWKS endpoint. The account

        is resolved from the `sub` claim against `accounts.auth_subject`, never from

        anything the client sends. Any `shopId` in a path is checked against that account

        before the query runs, and a shop belonging to another account returns 403 rather

        than 404, because the shop does exist and pretending otherwise helps nobody.

        '
...
  responses:
...
    ValidationFailed:
      description: 'The request was well formed but its contents were rejected. `errors` names the

        offending fields so the interface can mark them rather than showing one message.

        '
      content:
        application/problem+json:
          schema:
            allOf:
            - $ref: '#/components/schemas/Problem'
            - type: object
              properties:
                errors:
                  type: array
                  items:
                    type: object
                    required:
                    - field
                    - message
                    properties:
                      field:
                        type: string
                      message:
                        type: string
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details, with a stable machine-readable `code`.
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          description: Human readable. Wording may change.
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: 'The contract. Clients branch on this, never on `title` or `status` alone.

            '
          enum:
          - unauthenticated
          - token_expired
          - forbidden_shop
          - not_found
          - validation_failed
          - conflict
          - idempotency_key_conflict
          - rate_limited
          - shop_needs_reconnect
          - shop_region_unsupported
          - upstream_unavailable
          - internal_error
...
    AccountStatusOut:
      properties:
        account_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - active
          - suspended
        renewal:
          type: string
      type: object
      required:
      - account_id
      - status
      - renewal
...
    AdminAccountList:
      properties:
        accounts:
          items:
            $ref: '#/components/schemas/AdminAccount'
          type: array
      type: object
      required:
      - accounts
...
    AdminAuditList:
      properties:
        entries:
          items:
            $ref: '#/components/schemas/AdminAuditEntry'
          type: array
      type: object
      required:
      - entries
...
    AdminFigures:
      properties:
        currency:
          type: string
        accounts:
          type: integer
        accounts_active:
          type: integer
        signups_by_month:
          items:
            $ref: '#/components/schemas/MonthCount'
          type: array
        plans:
          items:
            $ref: '#/components/schemas/PlanFigure'
          type: array
        list_revenue_minor:
          type: integer
        note:
          type: string
      type: object
      required:
      - currency
      - accounts
      - accounts_active
      - signups_by_month
      - plans
      - list_revenue_minor
      - note
    AdminNoticeEmails:
      properties:
        counts:
          additionalProperties:
            type: integer
          type: object
        problems:
          items:
            $ref: '#/components/schemas/AdminEmailProblem'
          type: array
      type: object
      required:
      - counts
      - problems
...
    AdminShopList:
      properties:
        shops:
          items:
            $ref: '#/components/schemas/AdminShop'
          type: array
      type: object
      required:
      - shops
...